Simplify connector ctor to make it easier to integrate
