+++ cmd/baton-fullstory/main.go
@@ -3,17 +3,14 @@ package main
 import (
 	"context"
 	"fmt"
-	"net/http"
 	"os"
 
 	"github.com/conductorone/baton-sdk/pkg/connectorbuilder"
 	"github.com/conductorone/baton-sdk/pkg/field"
 	"github.com/conductorone/baton-sdk/pkg/types"
-	"github.com/conductorone/baton-sdk/pkg/uhttp"
 	"github.com/grpc-ecosystem/go-grpc-middleware/logging/zap/ctxzap"
 	"github.com/spf13/viper"
 	"go.uber.org/zap"
-	"golang.org/x/oauth2"
 
 	"github.com/conductorone/baton-fullstory/pkg/connector"
 	configschema "github.com/conductorone/baton-sdk/pkg/config"
@@ -39,36 +36,12 @@ func main() {
 	}
 }
 
-type CustomBasicAuth struct {
-	Token string
-}
-
-var _ uhttp.AuthCredentials = (*CustomBasicAuth)(nil)
-
-func (c *CustomBasicAuth) GetClient(ctx context.Context, options ...uhttp.Option) (*http.Client, error) {
-	httpClient, err := uhttp.NewClient(ctx, options...)
-	if err != nil {
-		return nil, fmt.Errorf("creating HTTP client failed: %w", err)
-	}
-
-	ctx = context.WithValue(ctx, oauth2.HTTPClient, httpClient)
-	ts := oauth2.StaticTokenSource(
-		&oauth2.Token{AccessToken: c.Token, TokenType: "basic"},
-	)
-	httpClient = oauth2.NewClient(ctx, ts)
-
-	return httpClient, nil
-}
-
 func getConnector(ctx context.Context, v *viper.Viper) (types.ConnectorServer, error) {
 	l := ctxzap.Extract(ctx)
 
-	var auth uhttp.AuthCredentials = &uhttp.NoAuth{}
-	if v.GetString("api-key") != "" {
-		auth = &CustomBasicAuth{Token: v.GetString("api-key")}
-	}
+	apiKey := v.GetString("api-key")
 
-	cb, err := connector.New(ctx, auth)
+	cb, err := connector.New(ctx, apiKey)
 	if err != nil {
 		l.Error("error creating connector", zap.Error(err))
 		return nil, err
